Move tests for nested test classes into regular test class
