+++ src/test/java/org/junitpioneer/jupiter/DefaultLocaleTests.java
@@ -124,6 +124,44 @@ class DefaultLocaleTests extends AbstractPioneerTestEngineTests {
 		}
 	}
 
+	@DisplayName("with nested classes")
+	@DefaultLocale(language = "en")
+	@Nested
+	class NestedDefaultLocaleTests extends AbstractPioneerTestEngineTests {
+
+		@Nested
+		@DisplayName("without DefaultLocale annotation")
+		class NestedClass {
+
+			@Test
+			@DisplayName("DefaultLocale should be set from enclosed class when it is not provided in nested")
+			public void shouldSetLocaleFromEnclosedClass() {
+				assertEquals(Locale.getDefault().getLanguage(), "en");
+			}
+		}
+
+		@Nested
+		@DefaultLocale(language = "de")
+		@DisplayName("with DefaultLocale annotation")
+		class AnnotatedNestedClass {
+
+			@Test
+			@DisplayName("DefaultLocale should be set from nested class when it is provided")
+			public void shouldSetLocaleFromNestedClass() {
+				assertEquals(Locale.getDefault().getLanguage(), "de");
+			}
+
+			@Test
+			@DefaultLocale(language = "ch")
+			@DisplayName("DefaultLocale should be set from method when it is provided")
+			public void shouldSetLocaleFromMethodOfNestedClass() {
+				assertEquals(Locale.getDefault().getLanguage(), "ch");
+			}
+		}
+
+	}
+
+
 	@Nested
 	@DisplayName("when configured incorrect")
 	class ConfigurationFailureTests {
